Inline action button

A button widget similar to shiny::actionButton() that can be included in an inline() wrapper.

Usage

inline_button(id, label, icon = NULL, meaning = label, accent = NULL)

Arguments

id

The input slot that will be used to access the value.

label

The text appearing within the button. This can be a character string or any other HTML, or NULL for no text (but then you will probably at least want an icon).

icon

An optional shiny::icon() which will appear to the left of the button.

meaning

A descriptive label, for people using assistive technology such as screen readers.

accent

A Bootstrap "accent" (such as "primary", "danger", etc.) that will be used to set the class of the button (such as "btn-primary", etc.), or NULL for the default ("btn-default"). See Bootstrap 5 buttons for all the options. If you provide multiple accents in a character vector, each one will be appended to "btn-" and added to the button.

Value

An inline widget to be included in an inline() wrapper.

See also

Examples

ui <- bslib::page_fixed(
    shiny::h1("A wonderful button"),
    inline("To update, please feel free to press the ",
        inline_button("mybutton",
            label = shiny::span(style = "font-style:italic", "button"),
            icon = shiny::icon("play"),
            meaning = "Update button", accent = "success"),
        "."
    )
)